- Deprecated old hotness algorithm - Small refactoring and adjustments
